requirements.txt
pandas==1.3.5
requests==2.26.0
gspread==4.0.1
google-auth==2.6.1